+++ dev/documentAll.sh
@@ -13,11 +13,16 @@ find "$REPO/xsl" -maxdepth 1 -name '*.xsl' -exec sh -c '
   file="$1"
   base=$(basename "$file")
   echo Generating HTML for $1...
-  java -jar "$2/utilities/saxon12he/saxon12he.jar" -a -warnings:silent -s:"$file" > "$2/xsl/$base.html"
+  java -jar "$2/utilities/saxonhe/saxonhe.jar" -a -warnings:silent -s:"$file" > "$2/xsl/$base.html"
 ' sh {} "$REPO" \;
 
 echo
 echo These files have inconsistencies that need to be addressed:
-find "$REPO/xsl" -maxdepth 1 -name \*.xsl.html -exec grep -l Inconsistencies {} \;
+find "$REPO" \
+  \( -name utilities -o -name xslstyle -o -name Crane-txt2xml \) -prune \
+  -name \*.html \
+  -o \( -name '*.xsl' \) \
+  -exec grep -l Inconsistencies {} \;
 echo End of list of files with inconsistencies
+
 echo Remember that the SXWN9040 errors are innocuous ... would love to avoid them but I haven\'t the time to update the 22-year-old files